% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celcius2fahrenheit.R
\name{celcius2fahrenheit}
\alias{celcius2fahrenheit}
\title{A function to convert a Celcius degree to a Farenheit degree}
\usage{
celcius2fahrenheit(temperature)
}
\arguments{
\item{temperature}{A Celcius degree numberic vector with one element}
}
\value{
A Fahrenheit degree number
}
\description{
A function to convert a Celcius degree to a Farenheit degree
}
\examples{
x <- "5"
celcius2fahrenheit(5)
}
