% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit2celcius.R
\name{fahrenheit2celcius}
\alias{fahrenheit2celcius}
\title{A function to convert a Farenheit degree to a Celcius degree}
\usage{
fahrenheit2celcius(temperature)
}
\arguments{
\item{temperature}{A Farenheit degree numberic vector with one element}
}
\value{
A Celcius degree number
}
\description{
A function to convert a Farenheit degree to a Celcius degree
}
\examples{
x <- 50
fahrenheit2celcius(x)
}
