% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit2kelvin.R
\name{fahrenheit2kelvin}
\alias{fahrenheit2kelvin}
\title{A function to convert a Farenheit degree to a Kelvin degree}
\usage{
fahrenheit2kelvin(temperature)
}
\arguments{
\item{temperature}{A Farenheit degree numberic vector with one element}
}
\value{
A Kelvin degree number
}
\description{
A function to convert a Farenheit degree to a Kelvin degree
}
\examples{
x <- 80.33
fahrenheit2kelvin(x)
}
