% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klevin2fahrenheit.R
\name{kelvin2fahrenheit}
\alias{kelvin2fahrenheit}
\title{A function to convert a Kelvin degree to a Farenheit degree}
\usage{
kelvin2fahrenheit(temperature)
}
\arguments{
\item{temperature}{A Kelvin degree numberic vector with one element}
}
\value{
A Fahrenheit degree number
}
\description{
A function to convert a Kelvin degree to a Farenheit degree
}
\examples{
x <- 300
kelvin2fahrenheit(x)
}
